fix(grid): optimize code
