+++ src/main.ts
@@ -7,8 +7,7 @@ import Icon from './icon';
 import Tag from './tag';
 import Notify from './notify';
 import Dropdown from './dropdown';
-import Row from './grid/row';
-import Col from './grid/col';
+import { Row, Col } from './grid';
 import CopyLink from './copy-link';
 import Radio from './radio';
 import Checkbox, { CheckboxGroup } from './checkbox';
